taxa: fix taxon selection
Translating a command-line taxon name to taxon entry meant
picking the first match. This could select a useless synonym
entry with no underlying taxa, and there seems to be quite
a lot of those in the data set -- perhaps for historical
reasons.

I suppose we should really walk through all matches, but
for now just pick the first accepted name.